Fix implicit type conversion warnings on MSVC

+++ src/GameLogic/Game/ApplicationData.h
@@ -27,7 +27,7 @@ public:
 	ApplicationData(int threadsCount);
 
 	void writeProfilingData();
-	void threadSaveProfileData(int threadIndex);
+	void threadSaveProfileData(size_t threadIndex);
 	void shutdownThreads();
 
 private:
